refactor: Moving GraphRepository interface to model
