+++ __tests__/unit/simple_graph_repository.test.ts
@@ -1,4 +1,4 @@
-import { SimpleGraphRepository } from "../../src/libs/engine/graph_repository/simple_graph_repository.class";
+import { SimpleGraphRepository } from "../../src/libs/engine/simple_graph_repository/simple_graph_repository.class";
 
 /**
  *  Tests the Simple Graph Repository.
