--- src/libs/engine/graph_repository/index.ts
@@ -1,4 +1,0 @@
-import { SimpleGraphRepository } from "./simple_graph_repository.class";
-import { GraphRepository } from "./graph_repository.interface";
-
-export { GraphRepository, SimpleGraphRepository };
