+++ src/libs/engine/index.ts
@@ -5,13 +5,11 @@ import {
 } from "@libs/engine/derivation_engine";
 import {
   SimpleGraphRepository,
-  GraphRepository,
-} from "@libs/engine/graph_repository";
+} from "@libs/engine/simple_graph_repository";
 
 export {
   PatternAnalysisEngine,
   DerivationEngine,
   DerivationRule,
   SimpleGraphRepository,
-  GraphRepository,
 };
