+++ src/libs/engine/simple_graph_repository/index.ts
@@ -1,0 +1,3 @@
+import { SimpleGraphRepository } from "./simple_graph_repository.class";
+
+export { SimpleGraphRepository };
