+++ src/libs/model/graph_repository/enums/edge_scope.enum.ts
