+++ src/libs/model/graph_repository/index.ts
@@ -1,0 +1,3 @@
+import { GraphRepository } from "@libs/model/graph_repository/graph_repository.interface";
+
+export { GraphRepository };
